# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"SendTransactionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"transactionHash": {
						"type": "string"
					},
					"fromAddress": {
						"type": "string"
					},
					"toAddress": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					},
					"tokenAddress": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"fromAddress",
					"toAddress",
					"value",
					"chainId",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendTransactionRequest": {
				"properties": {
					"fromAddress": {
						"type": "string"
					},
					"toAddress": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"tokenAddress": {
						"type": "string"
					},
					"gasLimit": {
						"type": "string"
					},
					"maxFeePerGas": {
						"type": "string"
					},
					"maxPriorityFeePerGas": {
						"type": "string"
					},
					"nonce": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fromAddress",
					"toAddress",
					"value",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenTransfer": {
				"properties": {
					"tokenAddress": {
						"type": "string"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"tokenSymbol": {
						"type": "string"
					},
					"tokenDecimals": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"tokenAddress",
					"from",
					"to",
					"value",
					"tokenSymbol",
					"tokenDecimals"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Transaction": {
				"properties": {
					"hash": {
						"type": "string"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"blockNumber": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"enum": [
							"success",
							"failed",
							"pending"
						]
					},
					"gasUsed": {
						"type": "string"
					},
					"gasPrice": {
						"type": "string"
					},
					"tokenTransfers": {
						"items": {
							"$ref": "#/components/schemas/TokenTransfer"
						},
						"type": "array"
					}
				},
				"required": [
					"hash",
					"from",
					"to",
					"value",
					"timestamp",
					"blockNumber",
					"status",
					"gasUsed",
					"gasPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionHistoryResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/Transaction"
						},
						"type": "array"
					},
					"walletAddress": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"transactions",
					"walletAddress",
					"chainId",
					"chainName",
					"page",
					"limit",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionHistoryRequest": {
				"properties": {
					"walletAddress": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"walletAddress",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GasPriceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"gasPrice": {
						"type": "string"
					},
					"maxFeePerGas": {
						"type": "string"
					},
					"maxPriorityFeePerGas": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"chainId",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionStatusResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"transactionHash": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"success",
							"failed",
							"pending"
						]
					},
					"blockNumber": {
						"type": "number",
						"format": "double"
					},
					"confirmations": {
						"type": "number",
						"format": "double"
					},
					"gasUsed": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"transactionHash",
					"status",
					"confirmations"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"app": {
				"type": "apiKey",
				"name": "X-App-ID",
				"in": "header",
				"description": "Application ID for API access"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT token for authentication"
			}
		}
	},
	"info": {
		"title": "crefy-connect-backend",
		"version": "1.0.0",
		"description": "Wallet authentication API with Express and TypeScript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jimleston Osoi"
		}
	},
	"paths": {
		"/transactions/send": {
			"post": {
				"operationId": "SendTransaction",
				"responses": {
					"200": {
						"description": "Transaction sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendTransactionResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"transactionHash": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
											"fromAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"toAddress": "0x742E6B6D8B6C4e8D8e8D8e8D8e8D8e8D8e8D8e8D",
											"value": "0.1",
											"chainId": 8453,
											"chainName": "Base"
										}
									}
								}
							}
						}
					}
				},
				"description": "Send transaction (native token or ERC20)",
				"tags": [
					"Transaction Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendTransactionRequest"
							}
						}
					}
				}
			}
		},
		"/transactions/history": {
			"post": {
				"operationId": "GetTransactionHistory",
				"responses": {
					"200": {
						"description": "Transaction history retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionHistoryResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"transactions": [
												{
													"hash": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
													"from": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
													"to": "0x742E6B6D8B6C4e8D8e8D8e8D8e8D8e8D8e8D8e8D",
													"value": "1000000000000000000",
													"timestamp": 1672531200,
													"blockNumber": 12345678,
													"status": "success",
													"gasUsed": "21000",
													"gasPrice": "20000000000"
												}
											],
											"walletAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"chainId": 8453,
											"chainName": "Base",
											"page": 1,
											"limit": 10,
											"total": 1
										}
									}
								}
							}
						}
					}
				},
				"description": "Get transaction history for a wallet",
				"tags": [
					"Transaction Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionHistoryRequest"
							}
						}
					}
				}
			}
		},
		"/transactions/gas-prices": {
			"get": {
				"operationId": "GetGasPrices",
				"responses": {
					"200": {
						"description": "Gas prices retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GasPriceResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"gasPrice": "30.5",
											"maxFeePerGas": "32.1",
											"maxPriorityFeePerGas": "1.5",
											"chainId": 8453,
											"chainName": "Base"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get current gas prices for a chain",
				"tags": [
					"Transaction Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/transactions/status": {
			"get": {
				"operationId": "GetTransactionStatus",
				"responses": {
					"200": {
						"description": "Transaction status retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionStatusResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"transactionHash": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
											"status": "success",
											"blockNumber": 12345678,
											"confirmations": 15,
											"gasUsed": "21000"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get transaction status",
				"tags": [
					"Transaction Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transactionHash",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/transactions/estimate-gas": {
			"post": {
				"operationId": "EstimateGas",
				"responses": {
					"200": {
						"description": "Gas estimated successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										},
										"gasEstimate": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"gasEstimate": "21000"
										}
									}
								}
							}
						}
					}
				},
				"description": "Estimate transaction gas",
				"tags": [
					"Transaction Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendTransactionRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}